style: 깃허브 코멘트 반영
