--- src/main/java/com/walletguardians/WalletGuardiansApiApplication.java
@@ -1,13 +1,0 @@
-package com.walletguardians;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class WalletGuardiansApiApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(WalletGuardiansApiApplication.class, args);
-	}
-
-}
