--- src/main/java/com/walletguardians/domain/user/dto/request/UserLoginRegister.java
@@ -1,9 +1,0 @@
-package com.walletguardians.domain.user.dto.request;
-
-import lombok.Getter;
-
-@Getter
-public class UserLoginRegister {
-  private String email;
-  private String password;
-}
