--- src/main/java/com/walletguardians/domain/user/entity/User.java
@@ -1,62 +1,0 @@
-package com.walletguardians.domain.user.entity;
-
-import com.walletguardians.domain.user.entity.auth.Role;
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.EnumType;
-import jakarta.persistence.Enumerated;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.Table;
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import org.springframework.security.crypto.password.PasswordEncoder;
-
-@Entity
-@Getter
-@Builder
-@AllArgsConstructor
-@Table(name = "USERS")
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class User {
-
-  @Id
-  @GeneratedValue(strategy = GenerationType.IDENTITY)
-  @Column(name = "user_id")
-  private Long id;
-
-  @Column(nullable = false)
-  private String title;
-
-  @Column(nullable = false)
-  private String name;
-
-  @Column(nullable = false, unique = true, length = 30)
-  private String email;
-
-  @Column(nullable = false)
-  private String password;
-
-  @Column(nullable = false, name = "defense-rate")
-  private float defenseRate;
-
-  @Column(nullable = false, name = "user-deleted")
-  private boolean userDeleted;
-
-  @Enumerated(EnumType.STRING)
-  private Role role;
-
-  //== 패스워드 암호화 ==//
-  public void encodePassword(PasswordEncoder passwordEncoder) {
-    this.password = passwordEncoder.encode(password);
-  }
-
-  public boolean isPasswordValid(PasswordEncoder passwordEncoder, String password) {
-    return passwordEncoder.matches(password, this.password);
-  }
-
-}
