--- src/main/java/com/walletguardians/domain/user/repository/UserRepository.java
@@ -1,11 +1,0 @@
-package com.walletguardians.domain.user.repository;
-
-import com.walletguardians.domain.user.entity.User;
-import java.util.Optional;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-public interface UserRepository extends JpaRepository<User, Long> {
-
-  Optional<User> findByEmail(String email);
-}
