--- src/main/java/com/walletguardians/global/auth/custom/CustomUserDetails.java
@@ -1,56 +1,0 @@
-package com.walletguardians.global.auth.custom;
-
-import com.walletguardians.domain.user.entity.User;
-import java.util.ArrayList;
-import java.util.Collection;
-import java.util.List;
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
-import org.springframework.security.core.userdetails.UserDetails;
-
-@Getter
-@RequiredArgsConstructor
-public class CustomUserDetails implements UserDetails {
-
-  private final User user;
-
-  @Override
-  public Collection<? extends GrantedAuthority> getAuthorities() {
-    List<GrantedAuthority> authorities = new ArrayList<>();
-    user.getRole().getRoles()
-        .forEach(role -> authorities.add(new SimpleGrantedAuthority(role)));
-    return authorities;
-  }
-
-  @Override
-  public String getPassword() {
-    return user.getPassword();
-  }
-
-  @Override
-  public String getUsername() {
-    return user.getEmail();
-  }
-
-  @Override
-  public boolean isAccountNonExpired() {
-    return true;
-  }
-
-  @Override
-  public boolean isAccountNonLocked() {
-    return true;
-  }
-
-  @Override
-  public boolean isCredentialsNonExpired() {
-    return true;
-  }
-
-  @Override
-  public boolean isEnabled() {
-    return true;
-  }
-}
