--- src/main/java/com/walletguardians/global/auth/custom/CustomUserDetailsService.java
@@ -1,24 +1,0 @@
-package com.walletguardians.global.auth.custom;
-
-import com.walletguardians.domain.user.entity.User;
-import com.walletguardians.domain.user.repository.UserRepository;
-import lombok.RequiredArgsConstructor;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.stereotype.Service;
-
-@Service
-@RequiredArgsConstructor
-public class CustomUserDetailsService implements UserDetailsService {
-
-  private final UserRepository userRepository;
-
-  @Override
-  public UserDetails loadUserByUsername(String email) throws UsernameNotFoundException {
-    User user = userRepository.findByEmail(email)
-        .orElseThrow(() -> new IllegalArgumentException(email + "에 해당하는 정보가 없습니다."));
-    return new CustomUserDetails(user);
-  }
-
-}
