--- src/main/java/com/walletguardians/global/auth/jwt/entity/RefreshToken.java
@@ -1,23 +1,0 @@
-package com.walletguardians.global.auth.jwt.entity;
-
-import jakarta.persistence.Entity;
-import jakarta.persistence.Id;
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Builder
-@Getter
-@AllArgsConstructor
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class RefreshToken {
-
-  @Id
-  private String userEmail;
-
-  private String refreshToken;
-
-}
