--- src/main/java/com/walletguardians/global/auth/jwt/repository/RefreshTokenRepository.java
@@ -1,9 +1,0 @@
-package com.walletguardians.global.auth.jwt.repository;
-
-import com.walletguardians.global.auth.jwt.entity.RefreshToken;
-import org.springframework.data.repository.CrudRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface RefreshTokenRepository extends CrudRepository<RefreshToken, String> {
-}
