+++ src/main/java/com/walletguardians/walletguardiansapi/WalletGuardiansApiApplication.java
@@ -1,0 +1,13 @@
+package com.walletguardians.walletguardiansapi;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class WalletGuardiansApiApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(WalletGuardiansApiApplication.class, args);
+	}
+
+}
