+++ src/main/java/com/walletguardians/walletguardiansapi/domain/user/dto/request/UserLoginRegister.java
@@ -1,0 +1,9 @@
+package com.walletguardians.walletguardiansapi.domain.user.dto.request;
+
+import lombok.Getter;
+
+@Getter
+public class UserLoginRegister {
+  private String email;
+  private String password;
+}
