+++ src/main/java/com/walletguardians/walletguardiansapi/domain/user/dto/request/UserRegisterRequest.java
@@ -1,0 +1,23 @@
+package com.walletguardians.walletguardiansapi.domain.user.dto.request;
+
+import com.walletguardians.walletguardiansapi.domain.user.entity.User;
+import com.walletguardians.walletguardiansapi.domain.user.entity.auth.Role;
+import lombok.Getter;
+
+@Getter
+public class UserRegisterRequest {
+  private String email;
+  private String password;
+  private String username;
+
+
+  public User toUserEntity() {
+    return User.builder()
+        .email(email)
+        .password(password)
+        .username(username)
+        .role(Role.USER)
+        .build();
+  }
+
+}
