+++ src/main/java/com/walletguardians/walletguardiansapi/domain/user/entity/User.java
@@ -1,0 +1,67 @@
+package com.walletguardians.walletguardiansapi.domain.user.entity;
+
+import com.walletguardians.walletguardiansapi.domain.user.entity.auth.Role;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.EnumType;
+import jakarta.persistence.Enumerated;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.Table;
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Builder.Default;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import org.hibernate.jdbc.Expectation.None;
+import org.springframework.security.crypto.password.PasswordEncoder;
+
+@Entity
+@Getter
+@Builder
+@AllArgsConstructor
+@Table(name = "USERS")
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class User {
+
+  @Id
+  @GeneratedValue(strategy = GenerationType.IDENTITY)
+  @Column(name = "user_id")
+  private Long id;
+
+  @Column(nullable = false)
+  private String username;
+
+  @Column(nullable = false, unique = true, length = 30)
+  private String email;
+
+  @Column(nullable = false)
+  private String password;
+
+  @Default
+  @Column(nullable = false)
+  private String title = "";
+
+  @Default
+  @Column(nullable = false, name = "defense-rate")
+  private float defenseRate = 0;
+
+  @Default
+  @Column(nullable = false, name = "user-deleted")
+  private boolean userDeleted = false;
+
+  @Enumerated(EnumType.STRING)
+  private Role role;
+
+  //== 패스워드 암호화 ==//
+  public void encodePassword(PasswordEncoder passwordEncoder) {
+    this.password = passwordEncoder.encode(password);
+  }
+
+  public boolean isPasswordValid(PasswordEncoder passwordEncoder, String password) {
+    return passwordEncoder.matches(password, this.password);
+  }
+
+}
