+++ src/main/java/com/walletguardians/walletguardiansapi/domain/user/repository/UserRepository.java
@@ -1,0 +1,11 @@
+package com.walletguardians.walletguardiansapi.domain.user.repository;
+
+import com.walletguardians.walletguardiansapi.domain.user.entity.User;
+import java.util.Optional;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface UserRepository extends JpaRepository<User, Long> {
+
+  Optional<User> findByEmail(String email);
+
+}
