+++ src/main/java/com/walletguardians/walletguardiansapi/domain/user/service/UserService.java
@@ -1,0 +1,26 @@
+package com.walletguardians.walletguardiansapi.domain.user.service;
+
+import com.walletguardians.walletguardiansapi.domain.user.entity.User;
+import com.walletguardians.walletguardiansapi.domain.user.repository.UserRepository;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+@Service
+@RequiredArgsConstructor
+public class UserService {
+
+  private final UserRepository userRepository;
+
+  public User findByUserId(Long userId) {
+    return userRepository.findById(userId)
+        .orElseThrow(() -> new IllegalArgumentException("해당하는 회원 정보가 없습니다."));
+  }
+
+  @Transactional(readOnly = true)
+  public User findUserByEmail(String email) {
+    return userRepository.findByEmail(email)
+        .orElseThrow(() -> new IllegalArgumentException("해당하는 이메일 정보가 없습니다."));
+  }
+
+}
