+++ src/main/java/com/walletguardians/walletguardiansapi/global/auth/jwt/dto/TokenDto.java
@@ -1,0 +1,15 @@
+package com.walletguardians.walletguardiansapi.global.auth.jwt.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor
+@AllArgsConstructor
+public class TokenDto {
+
+  private String accessToken;
+  private String refreshToken;
+
+}
